refactor(webapp): split WC live boot wiring
