+++ packages/webapp/src/ui/wc/wc-live.ts
@@ -11,34 +11,10 @@ import type { BrowserAPI, CDPTransport } from '../../cdp/index.js';
 import { isFeatureEnabled } from '../../core/feature-flags.js';
 import { SessionStore as AgentSessionStore } from '../../core/session.js';
 import { installPageStorageSync } from '../../kernel/page-storage-sync.js';
-import {
-  SYNC_FS_NEED_NONCE_MSG,
-  SYNC_FS_NONCE_MSG,
-  type SyncFsNeedNonceMsg,
-  type SyncFsNonce,
-  type SyncFsNonceMsg,
-} from '../../kernel/realm/sync-fs-wire.js';
 import type { RemoteTerminalView } from '../../kernel/remote-terminal-view.js';
 import { spawnKernelWorker } from '../../kernel/spawn.js';
-import {
-  getAccounts,
-  getOAuthAccountInfo,
-  getProviderConfig,
-  resolveCurrentModel,
-  resolveModelById,
-} from '../../providers/account-store.js';
 import { SessionStore as UiSessionStore } from '../../scoops/chat-session-store.js';
-import type { LickEvent } from '../../scoops/lick-manager.js';
-import { getFollowerTrayRuntimeStatus } from '../../scoops/tray-follower-status.js';
-import { getLeaderTrayRuntimeStatus } from '../../scoops/tray-leader.js';
-import {
-  hasStoredTrayJoinUrl,
-  normalizeTrayWorkerBaseUrl,
-  parseTrayJoinUrlValue,
-  TRAY_WORKER_STORAGE_KEY,
-} from '../../scoops/tray-runtime-config.js';
-import type { RegisteredScoop, ThinkingLevel } from '../../scoops/types.js';
-import { getConnectedFollowers } from '../../shell/supplemental-commands/host-command.js';
+import type { RegisteredScoop } from '../../scoops/types.js';
 import { registerTranscriptExportService } from '../../transcript/export-provider.js';
 import { DefaultTranscriptExportService } from '../../transcript/export-service.js';
 import { readSnapshot, writeSnapshot } from '../../transcript/snapshot-store.js';
@@ -49,711 +25,41 @@ import {
 } from '../boot/setup-preload-error-reload.js';
 import { setupStandalonePrelude } from '../boot/setup-standalone-prelude.js';
 import type { BootStageLogger } from '../boot/types.js';
-import { type DipInstance, disposeDips, hydrateDips } from '../dip.js';
-import { isLickChannel } from '../lick-channels.js';
-import {
-  OffscreenClient,
-  type OffscreenClientCallbacks,
-  type ScoopBusyPhase,
-} from '../offscreen-client.js';
+import { OffscreenClient } from '../offscreen-client.js';
 import type { UiRuntimeMode } from '../runtime-mode.js';
-import type { ChatMessage } from '../types.js';
-import { notifyLeaderLocalModelStateChanged } from './leader-model-events.js';
+import type { WcChatController } from './wc-chat-controller.js';
 import {
-  LEADER_BROADCAST_SNAPSHOT_EVENT,
-  LEADER_RUN_NEW_SESSION_EVENT,
-  type LeaderRunNewSessionDetail,
-} from './leader-session-events.js';
-import { WcChatController } from './wc-chat-controller.js';
-import {
-  effortOverrideForAgent,
-  metaThinkingForScoop,
-  thinkingLevelForAgent,
-} from './wc-follower-model-surface.js';
-import type { MonitorTrayInfo } from './wc-monitor.js';
-import { scoopColor } from './wc-scoop-color.js';
-
-export { scoopColor } from './wc-scoop-color.js';
-
-import type { FrozenSession } from '../session-freezer.js';
-import {
-  enrichFreezerIcons,
-  FREEZER_TINT,
-  type FrozenSessionIndexEntry,
-  readFreezerEntries,
-  readFreezerIndexState,
-  rebuildFreezerIndexFromArchives,
-  renderFreezerCards,
-  SESSIONS_INDEX_PATH,
-  thawFrozenSession,
-} from './wc-freezer.js';
-import {
-  applyShellContext,
-  mountWcShell,
-  type SwitcherScoop,
-  submittedSteer,
-  submittedText,
-  type WcShellRefs,
-} from './wc-shell.js';
+  createWcLiveCallbacks,
+  type LickBackpressureState,
+  toSwitcherScoops,
+  type WcLiveWiring,
+} from './wc-live-callbacks.js';
+import { wireWcComposer } from './wc-live-composer.js';
+import { createWcController, type WelcomeInterceptHolder } from './wc-live-controller.js';
+import { wireFreezerRail } from './wc-live-freezer.js';
+import { createWcMonitorDeps } from './wc-live-monitor-deps.js';
+import { setupSyncFsBootNonce } from './wc-live-sync-fs.js';
+import { applyThreadContext } from './wc-live-thinking-hydration.js';
+import { mountWcShell, type WcShellRefs } from './wc-shell.js';
 import { createWorkbenchActivator, type WorkbenchActivator } from './wc-workbench.js';
 
 export {
+  createWcLiveCallbacks,
+  type LickBackpressureState,
+  type ScoopStatus,
+  toSwitcherScoops,
+  type WcLiveWiring,
+} from './wc-live-callbacks.js';
+export type { WelcomeInterceptHolder } from './wc-live-controller.js';
+export { parseProcStatLine } from './wc-live-monitor-deps.js';
+export {
+  applyLeaderLocalThinkingChange,
   effortOverrideForAgent,
   metaThinkingForScoop,
+  shouldSkipSessionHydration,
   thinkingLevelForAgent,
-} from './wc-follower-model-surface.js';
-
-/** Persist a leader-local thinking change and notify followers only after its ack. */
-export async function applyLeaderLocalThinkingChange(
-  client: Pick<OffscreenClient, 'setScoopThinkingLevel'>,
-  scoopJid: string,
-  level: ThinkingLevel | undefined,
-  effortOverride?: string,
-  notify: () => void = notifyLeaderLocalModelStateChanged
-): Promise<boolean> {
-  const applied = await client.setScoopThinkingLevel(scoopJid, level, effortOverride);
-  if (applied) notify();
-  return applied;
-}
-
-/**
- * Maps proc-mount.ts's single-letter process state code back to the word
- * the monitor UI checks for (wc-monitor.ts: `proc.status === 'running'`).
- * Unrecognized/missing letters map to `'unknown'` rather than throwing —
- * a process that exits mid-read (ENOENT on the next file) is handled by the
- * caller's try/catch, not here.
- */
-const PROC_STATE_LETTER_TO_WORD: Record<string, string> = {
-  R: 'running',
-  S: 'pending',
-  Z: 'exited',
-  K: 'killed',
-};
-
-/**
- * Parses a `/proc/<pid>/stat` line — proc-mount.ts's `renderStat()` output:
- * `"<pid> (<kind>) <stateLetter> <ppid> <exitCode> <startedAt> <finishedAt>"`
- * — and returns the process status word for field 3 (the state letter).
- *
- * Extracted as a standalone pure function (rather than inlined in
- * `getProcesses` below) specifically so this parsing/mapping logic — the
- * fragile part of the fix for the "processes never show as active" bug
- * (the state letter must land in the right array index, and the letter
- * must map to the exact word `wc-monitor.ts` compares against) — has real
- * unit test coverage without needing to mock the VFS/OffscreenClient chain
- * `getProcesses` depends on.
- *
- * Historical note: this used to read the *verbose* `/proc/<pid>/status`
- * dump (`Name:\t...\nState:\tR (running)\n...\nCmdline:\t...`) and pass the
- * whole multi-line blob through as `status` — `wc-monitor.ts`'s
- * `proc.status === 'running'` check could never match a multi-line string,
- * so the active/ended dot was always grey regardless of real process state.
- */
-export function parseProcStatLine(statLine: string): string {
-  const stateLetter = statLine.trim().split(' ')[2] ?? '';
-  return PROC_STATE_LETTER_TO_WORD[stateLetter] ?? 'unknown';
-}
-
-function getTrayMonitorInfo(storage: Pick<Storage, 'getItem'>): MonitorTrayInfo {
-  const follower = getFollowerTrayRuntimeStatus();
-  if (follower.state !== 'inactive') {
-    return {
-      role: 'follower',
-      state: follower.state,
-      joinUrl: follower.joinUrl,
-      sessionId: follower.trayId,
-      workerBaseUrl: parseTrayJoinUrlValue(follower.joinUrl)?.workerBaseUrl ?? null,
-      stalled: follower.stalled,
-    };
-  }
-
-  const leader = getLeaderTrayRuntimeStatus();
-  if (leader.state !== 'inactive') {
-    return {
-      role: 'leader',
-      state: leader.state,
-      joinUrl: leader.session?.joinUrl,
-      sessionId: leader.session?.trayId,
-      workerBaseUrl: leader.session?.workerBaseUrl,
-    };
-  }
-
-  let workerBaseUrl: string | null = null;
-  try {
-    workerBaseUrl = normalizeTrayWorkerBaseUrl(storage.getItem(TRAY_WORKER_STORAGE_KEY));
-  } catch {
-    // Storage can be unavailable in a sandboxed page; inactive tray display remains usable.
-  }
-  return { role: 'standalone', state: 'inactive', workerBaseUrl };
-}
-
-/**
- * Whether stale-session hydration should be skipped: true when the page is a
- * tray follower (stored join URL or cherry embed) or when a non-cone URL
- * context is pending. In all these cases the leader's snapshot replaces local
- * history, so flashing IndexedDB content first would cause flicker.
- */
-export function shouldSkipSessionHydration(
-  pendingUrlContext: string | null | undefined,
-  win: { location: { href: string }; localStorage: Storage }
-): boolean {
-  if (pendingUrlContext != null && pendingUrlContext !== 'cone') return true;
-  if (new URL(win.location.href).searchParams.get('cherry') === '1') return true;
-  return hasStoredTrayJoinUrl(win.localStorage);
-}
-
-/** Scoop runtime status, as broadcast by `onStatusChange`. */
-export type ScoopStatus = 'initializing' | 'ready' | 'processing' | 'error';
-
-export interface LickBackpressureState {
-  count: number;
-  waitingMs: number;
-}
-
-/** Chip eye state for a scoop status: errored scoops get the dead look. */
-function eyesFor(status: ScoopStatus | undefined): SwitcherScoop['eyes'] {
-  if (status === 'error') return 'dead';
-  if (status === 'initializing') return 'none';
-  return 'open';
-}
-
-/**
- * Map runtime lifecycle onto tab glyph state: processing → working,
- * error → broken, initializing → initializing, and ready → idle. A scoop
- * missing its first status broadcast is also idle because no work or failure
- * has been observed yet.
- */
-function stateFor(status: ScoopStatus | undefined): NonNullable<SwitcherScoop['state']> {
-  switch (status) {
-    case 'processing':
-      return 'working';
-    case 'error':
-      return 'broken';
-    case 'initializing':
-      return 'initializing';
-    case 'ready':
-    case undefined:
-      return 'idle';
-  }
-  status satisfies never;
-  return 'idle';
-}
-
-/** Map registered scoops onto switcher tab descriptors (cone first). */
-export function toSwitcherScoops(
-  scoops: readonly RegisteredScoop[],
-  statuses?: ReadonlyMap<string, ScoopStatus>,
-  fills?: ReadonlyMap<string, number>,
-  phases?: ReadonlyMap<string, ScoopBusyPhase>,
-  awaitingJid?: string | null
-): SwitcherScoop[] {
-  return [...scoops]
-    .sort((a, b) => Number(b.isCone) - Number(a.isCone))
-    .map((scoop) => {
-      const fill = fills?.get(scoop.jid);
-      const status = statuses?.get(scoop.jid);
-      return {
-        key: scoop.jid,
-        type: scoop.isCone ? 'cone' : 'scoop',
-        color: scoopColor(scoop),
-        label: scoop.isCone ? 'sliccy' : scoop.name,
-        eyes: eyesFor(status),
-        state: stateFor(status),
-        // The chip pupils dilate with context fullness (pill `fill` 0-100).
-        fill: typeof fill === 'number' ? Math.round(fill * 100) : undefined,
-        // Only a processing scoop has a busy phase; a stale entry for one that
-        // has since gone idle must not shape a pin that is no longer painted.
-        phase: status === 'processing' ? phases?.get(scoop.jid) : undefined,
-        // The scoop that just finished a turn is waiting on YOU, not idling.
-        awaiting: awaitingJid === scoop.jid || undefined,
-      };
-    });
-}
-
-export interface WcLiveWiring {
-  refs: WcShellRefs;
-  /** Live per-scoop status, written by the callbacks on every broadcast. */
-  statuses: Map<string, ScoopStatus>;
-  /** Per-scoop context-window fill (0..1), refreshed by the stats poller. */
-  fills: Map<string, number>;
-  /**
-   * Per-scoop busy phase (`tool` while a tool call is in flight). Written by
-   * `onScoopPhaseChange` for every scoop, selected or not; only read while a
-   * scoop is processing.
-   */
-  phases: Map<string, ScoopBusyPhase>;
-  /** Latest sustained lick-deferral state for each still-registered scoop. */
-  lickBackpressure: Map<string, LickBackpressureState>;
-  /**
-   * URL-restored boot context (the thread's `ctx` param) still awaiting
-   * routing. While set, the boot auto-select targets it instead of the cone
-   * (`scoop:<name>` selects that scoop; `freezer:<file>` suppresses selection
-   * until the host thaws it). Cleared once routed.
-   */
-  pendingUrlContext: string | null;
-  /**
-   * Most recent activity snippet per scoop jid (latest incoming message,
-   * user input, or finished reply) — the raw material for the hover-tooltip
-   * summaries and the `attention` (blinking eyes) bookkeeping.
-   */
-  lastActivity: Map<string, string>;
-  /**
-   * The scoop whose turn just ended and which is now waiting on the user —
-   * its avatar makes eye contact with the composer and eventually drowses.
-   * Set by the turn-finished hook, cleared the moment it works again.
-   */
-  awaitingInput?: string | null;
-  getController(): WcChatController | null;
-  getClient(): OffscreenClient | null;
-  getSelected(): RegisteredScoop | null;
-  selectScoop(scoop: RegisteredScoop): void;
-  /** Notify the tray leader after the rendered lifecycle transition gate fires. */
-  notifyScoopStateChanged?(): void;
-  /**
-   * Rebuild the switcher row from live state. Assigned BY
-   * `createWcLiveCallbacks` (the other direction from the hooks above) so
-   * wiring outside the callbacks — the turn-finished hook, the submit
-   * listener — can refresh without duplicating the descriptor mapping.
-   */
-  refreshScoops?(): void;
-  /** Fired once the kernel reports ready (late wiring re-runs boot reads). */
-  notifyReady?(): void;
-}
-
-/**
- * Kernel callbacks for the WC live shell. Pure factory over the wiring
- * handle so tests can drive it with fakes — no worker required.
- */
-export function createWcLiveCallbacks(wiring: WcLiveWiring): OffscreenClientCallbacks {
-  const refreshScoops = (): void => {
-    const client = wiring.getClient();
-    if (client) {
-      wiring.refs.switcher.scoops = toSwitcherScoops(
-        client.getScoops(),
-        wiring.statuses,
-        wiring.fills,
-        wiring.phases,
-        wiring.awaitingInput
-      );
-    }
-  };
-  wiring.refreshScoops = refreshScoops;
-  /** Read-only frozen-session view — selection is intentionally empty there. */
-  const viewingFrozen = (): boolean =>
-    (wiring.refs.thread.getAttribute('context') ?? '').startsWith('freezer:');
-  /**
-   * Select the boot target when nothing is selected. The first state snapshot
-   * can land BEFORE the cone finishes restoring (e.g. right after a clear +
-   * reload) — onReady then has no cone to select, and a restored cone only
-   * ever appears via later scoop-list updates (no scoop-created event). The
-   * frozen view is exempt: its empty selection is deliberate. A URL-restored
-   * context redirects the default: `scoop:<name>` selects that scoop,
-   * `freezer:<file>` keeps the selection empty for the host's thaw routing.
-   */
-  const ensureSelection = (): void => {
-    if (wiring.getSelected() || viewingFrozen()) return;
-    const scoops = wiring.getClient()?.getScoops() ?? [];
-    const pending = wiring.pendingUrlContext;
-    if (pending?.startsWith('freezer:')) return;
-    if (pending?.startsWith('scoop:')) {
-      const name = pending.slice('scoop:'.length);
-      const scoop = scoops.find((s) => !s.isCone && s.name === name);
-      if (scoop) {
-        wiring.pendingUrlContext = null;
-        wiring.selectScoop(scoop);
-        return;
-      }
-    }
-    const cone = scoops.find((s) => s.isCone);
-    if (cone) {
-      // The URL scoop is gone (dropped since) — the cone is the live truth.
-      wiring.pendingUrlContext = null;
-      wiring.selectScoop(cone);
-    }
-  };
-  return {
-    onStatusChange: (jid, status) => {
-      const previous = wiring.statuses.get(jid);
-      const next = status as ScoopStatus;
-      wiring.statuses.set(jid, next);
-      // Back at work (or broken): it is no longer waiting on the user.
-      if (next !== 'ready' && wiring.awaitingInput === jid) wiring.awaitingInput = null;
-      // The tabs setter rebuilds the Light DOM row and reflows it. Meaningful
-      // rendered transitions are worth that cost for the small runtime roster;
-      // suppress duplicate churn that would needlessly interrupt hover or focus.
-      if (eyesFor(previous) !== eyesFor(next) || stateFor(previous) !== stateFor(next)) {
-        refreshScoops();
-        wiring.notifyScoopStateChanged?.();
-      }
-      if (wiring.getSelected()?.jid !== jid) return;
-      wiring.getController()?.setProcessing(status === 'processing');
-    },
-    onScoopCreated: (scoop) => {
-      refreshScoops();
-      // A pending URL context owns the boot selection — don't steal it.
-      if (!wiring.getSelected() && !viewingFrozen() && !wiring.pendingUrlContext) {
-        wiring.selectScoop(scoop);
-      }
-    },
-    onScoopListUpdate: (scoops) => {
-      const registered = new Set(scoops.map((scoop) => scoop.jid));
-      for (const jid of wiring.lickBackpressure.keys()) {
-        if (!registered.has(jid)) wiring.lickBackpressure.delete(jid);
-      }
-      refreshScoops();
-      ensureSelection();
-    },
-    onScoopActivity: (jid) => {
-      // Agent-event ping for ANY scoop (selected or not): keep the navbar
-      // eyes on whichever scoop is actively streaming. Mirrors the
-      // `attention` write in `onIncomingMessage`; no thread routing.
-      wiring.refs.switcher.setAttribute('attention', jid);
-    },
-    onScoopPhaseChange: (jid, phase) => {
-      // Busy detail for ANY scoop: the tab's centre pin goes square while the
-      // model thinks and circular while a tool runs, the same vocabulary the
-      // composer's send button spends on motion. Deduped upstream, so this
-      // only repaints on an actual crossing.
-      wiring.phases.set(jid, phase);
-      refreshScoops();
-      // Followers render the same distinction from `ScoopSummary.state`, so a
-      // phase crossing is a wire-visible change — without this it would only
-      // reach them on the 5s list poll.
-      wiring.notifyScoopStateChanged?.();
-    },
-    onIncomingMessage: (jid, message) => {
-      // Most-recent-activity tracking: the scoop that just received a message
-      // wears the blinking navbar eyes (the switcher's `attention` chip) and
-      // the snippet feeds the hover-tooltip summary.
-      wiring.refs.switcher.setAttribute('attention', jid);
-      wiring.lastActivity.set(jid, String(message.content ?? '').slice(0, 600));
-      if (wiring.getSelected()?.jid !== jid) return;
-      if (message.channel !== 'web' && isLickChannel(message.channel)) {
-        wiring
-          .getController()
-          ?.addLickMessage(
-            message.id,
-            message.content,
-            message.channel,
-            new Date(message.timestamp).getTime(),
-            message.lickId
-          );
-      }
-    },
-    onLickBackpressure: (jid, info) => {
-      if (info.count <= 0) wiring.lickBackpressure.delete(jid);
-      else wiring.lickBackpressure.set(jid, info);
-      const selected = wiring.getSelected();
-      if (selected?.jid !== jid) return;
-      const scoopName = selected.isCone ? 'cone' : selected.name;
-      wiring.getController()?.setLickBackpressure(info.count, info.waitingMs, scoopName);
-    },
-    onMessageUpdate: (jid, update) => {
-      // Live flip of an actionable lick card's state (sudo-request settled).
-      // Only the selected scoop's thread is mounted, so a non-selected update
-      // is a no-op here — the persisted lickState rehydrates it on next load.
-      if (wiring.getSelected()?.jid !== jid) return;
-      if (update.lickId && update.lickState) {
-        wiring.getController()?.updateLickState(update.lickId, update.lickState);
-      }
-    },
-    onScoopMessagesReplaced: (jid, messages) => {
-      if (wiring.getSelected()?.jid !== jid) return;
-      wiring.getController()?.loadMessages(messages as unknown as ChatMessage[]);
-    },
-    onReady: () => {
-      refreshScoops();
-      ensureSelection();
-      wiring.notifyReady?.();
-    },
-  };
-}
-
-/** Point the thread chrome at a scoop (context label + accent hue + model). */
-async function applyThreadContext(refs: WcShellRefs, scoop: RegisteredScoop): Promise<void> {
-  refs.thread.setAttribute('context', scoop.isCone ? 'cone' : `scoop:${scoop.name}`);
-  refs.thread.setAttribute('accent', scoopColor(scoop));
-  refs.switcher.setAttribute('active', scoop.jid);
-  // The whole frame changes mood with the selection: Caramel Sugar Glass for
-  // the cone, swirling pastels + the scoop's accent for scoops.
-  applyShellContext(
-    refs,
-    scoop.isCone ? { kind: 'cone' } : { kind: 'scoop', accent: scoopColor(scoop) }
-  );
-  const lockedEffort = localStorage.getItem('slicc_locked_effort_level');
-  refs.composerMeta.setAttribute(
-    'thinking',
-    metaThinkingForScoop(
-      (lockedEffort ?? scoop.config?.thinkingLevel) as ThinkingLevel | undefined,
-      scoop.config?.effortOverride
-    )
-  );
-  try {
-    const { resolveCurrentModel, resolveModelById } = await import('../provider-settings.js');
-    const modelId = scoop.config?.modelId;
-    const model = modelId ? resolveModelById(modelId) : resolveCurrentModel();
-    refs.composerMeta.setAttribute('model', model.name ?? model.id);
-    // The thinking-effort pill only shows for a reasoning-capable model.
-    // When locked, hide the pill entirely so the user cannot change it.
-    refs.composerMeta.toggleAttribute(
-      'no-thinking',
-      (model as { reasoning?: boolean }).reasoning !== true || !!lockedEffort
-    );
-  } catch {
-    // Model display is informational; never block scoop selection on it.
-  }
-}
-
-interface FreezerRailDeps {
-  refs: WcShellRefs;
-  openVfs(): Promise<WcPageVfs>;
-  client: OffscreenClient;
-  getController(): WcChatController | null;
-  getSelected(): RegisteredScoop | null;
-  selectScoop(scoop: RegisteredScoop): void;
-  clearSelection(): void;
-  log: BootStageLogger;
-}
-
-/** Handles returned by {@link wireFreezerRail} for boot + URL routing. */
-interface FreezerRailHandles {
-  /** Re-read the frozen-session index and re-render the rail cards. */
-  refreshFreezer(): void;
-  /** Thaw a frozen session by archive filename (URL `ctx=freezer:<file>`). */
-  openFrozen(slug: string): Promise<void>;
-  /**
-   * Returns the `sessionId` of the currently viewed frozen session, or
-   * `null` when the thread is showing an active scoop rather than a thaw.
-   */
-  getViewedFrozenSessionId(): string | null;
-}
-
-/**
- * Freezer rail behavior: frozen sessions thaw read-only into the thread,
- * and the "New +" gestures freeze/clear the cone (click = LLM-enriched
- * freeze, double-click = quick-freeze, long-press = erase). Returns the
- * card-refresh function for the post-boot initial fill plus the by-slug
- * thaw used by URL routing.
- */
-// biome-ignore lint/complexity/noExcessiveLinesPerFunction: freezer wiring is sequential (rail refresh + new-session freeze + thaw + follower relay); extracting more helpers would obscure the order
-function wireFreezerRail(deps: FreezerRailDeps): FreezerRailHandles {
-  const { refs, openVfs, client, getController, getSelected, clearSelection, log } = deps;
-  let frozenEntries: FrozenSessionIndexEntry[] = [];
-  /** sessionId of the currently-viewed frozen session, or null for live view. */
-  let currentFrozenSessionId: string | null = null;
-  /** Wrap selectScoop to clear frozen-view tracking when returning to a live scoop. */
-  const selectScoop = (scoop: Parameters<FreezerRailDeps['selectScoop']>[0]): void => {
-    currentFrozenSessionId = null;
-    deps.selectScoop(scoop);
-  };
-
-  // Monotonic guard: boot fires several overlapping refreshes (attach-time,
-  // scoop-list ready, kernel-worker-ready) and a lost early RPC fails LATE —
-  // only the newest call may touch the rail, and faults never wipe it.
-  let refreshSeq = 0;
-  let iconEnriching = false;
-  const refreshFreezer = (): void => {
-    const seq = ++refreshSeq;
-    void openVfs()
-      .then(async ({ reader, writer }) => {
-        let entries = await readFreezerEntries(reader);
-        if (entries === null) {
-          // Faults keep the rail; a CORRUPT index (e.g. truncated by a
-          // reload that killed the worker mid-write) self-heals from the
-          // archives — they are the ground truth.
-          const state = await readFreezerIndexState(reader);
-          if (state.kind !== 'corrupt') return;
-          log.warn('WC freezer index corrupt — rebuilding from archives');
-          entries = await rebuildFreezerIndexFromArchives(reader);
-          if (entries.length === 0) return;
-          await writer.writeFile(SESSIONS_INDEX_PATH, JSON.stringify(entries, null, 2));
-        }
-        if (seq !== refreshSeq) return;
-        frozenEntries = entries;
-        renderFreezerCards(refs.freezer, entries);
-        // Backfill LLM-picked rail icons for icon-less entries (legacy /
-        // quick-frozen). Fire-and-forget and single-flight — refreshes
-        // re-fire on kernel-ready.
-        if (!iconEnriching && entries.some((e) => !e.icon && !e.pendingEnrichment)) {
-          iconEnriching = true;
-          void import('../../providers/quick-llm.js')
-            .then(({ pickLucideIcon }) =>
-              enrichFreezerIcons({
-                reader,
-                writer,
-                freezer: refs.freezer,
-                entries,
-                pickIcon: (subject) => pickLucideIcon({ subject }),
-              })
-            )
-            .catch((err) => log.warn('WC freezer icon enrichment failed', err))
-            .finally(() => {
-              iconEnriching = false;
-            });
-        }
-      })
-      .catch((err) => log.error('WC freezer refresh failed', err));
-  };
-
-  // One new-session at a time: a re-click while the freeze is in flight used
-  // to run a SECOND freeze (the stuck spinner invited it), leaving duplicate
-  // archives seconds apart. The spinner is set here and always cleared —
-  // the library only enters the busy state, it never exits it on its own.
-  let newSessionInFlight = false;
-  const freezerNew = (): HTMLElement | null => refs.freezer.querySelector('slicc-freezer-new');
-  const runNewSession = (action: 'save' | 'skip' | 'erase'): void => {
-    if (newSessionInFlight) return;
-    newSessionInFlight = true;
-    freezerNew()?.setAttribute('busy', '');
-    void (async () => {
-      try {
-        // The freezer itself skips empty/short sessions (MIN_MESSAGES_TO_FREEZE),
-        // so save/skip can always run it; erase intentionally never archives.
-        const { writer } = await openVfs();
-        const { resetNewSessionTmp, runNewSessionFreeze, runNewSessionFreezeQuick } = await import(
-          '../new-session.js'
-        );
-        if (action !== 'erase') {
-          // Capture a sanitized snapshot AFTER the Markdown archive write,
-          // BEFORE clearing. Non-blocking: failure marks the entry but never
-          // blocks New Session. No raw fallback on error.
-          const captureCompleteSnapshot = async (frozen: FrozenSession): Promise<void> => {
-            const { getTranscriptExportService } = await import(
-              '../../transcript/export-provider.js'
-            );
-            const svc = getTranscriptExportService();
-            await svc.captureFrozen({
-              sessionId: frozen.sessionId ?? frozen.archive.id,
-              title: frozen.archive.title,
-              frozenAt: frozen.archive.frozenAt,
-              createdAt: frozen.archive.createdAt,
-              updatedAt: frozen.archive.updatedAt,
-            });
-          };
-          if (action === 'save') {
-            // Write-first + race: the durable archive lands before any LLM
-            // call, and this resolves at min(LLM-done, race window) so the
-            // chat clears even if the provider is hung. The race timer drives
-            // the spinner's progress ring; background enrichment (timer-won)
-            // refreshes the rail when the late rename/icon land.
-            await runNewSessionFreeze({
-              vfs: writer,
-              agenticMemorySpawn: (options) => client.spawnAgent(options),
-              captureCompleteSnapshot,
-              onProgress: (fraction) => {
-                const el = freezerNew();
-                if (!el) return;
-                if (fraction === null) el.removeAttribute('progress');
-                else el.setAttribute('progress', String(fraction));
-              },
-              onBackgroundEnriched: () => refreshFreezer(),
-            });
-          } else {
-            // Quick-freeze: durable archive with heuristic title, no LLM
-            // enrichment (the entry stays pending-*.md forever).
-            await runNewSessionFreezeQuick({ vfs: writer, captureCompleteSnapshot });
-          }
-        }
-        await resetNewSessionTmp(writer);
-        await client.clearAllMessages();
-        // Clear the thread directly — re-selection only *requests* a replay,
-        // and the worker no-ops the reply for a (now) empty history, which
-        // left the old conversation on screen until the next reload.
-        getController()?.loadMessages([]);
-        // Tell wc-tray to broadcast the cleared snapshot to every connected
-        // follower — `clearAllMessages` emits no agent event, so already-
-        // connected followers would otherwise keep the stale chat until they
-        // reconnect or re-request a snapshot.
-        window.dispatchEvent(new CustomEvent(LEADER_BROADCAST_SNAPSHOT_EVENT));
-        // Re-arm the dictation priming note: the next dictated turn in the
-        // fresh session must carry the one-time TTS context note again.
-        void import('../../speech/dictation-priming.js')
-          .then(({ resetDictationPriming }) => resetDictationPriming())
-          .catch(() => undefined);
-        refreshFreezer();
-        const cone = client.getScoops().find((s) => s.isCone);
-        if (cone) selectScoop(cone);
-      } catch (err) {
-        log.error('WC new session failed', err);
-      } finally {
-        newSessionInFlight = false;
-        const el = freezerNew();
-        el?.removeAttribute('busy');
-        el?.removeAttribute('progress');
-      }
-    })();
-  };
-  for (const action of ['save', 'skip', 'erase'] as const) {
-    refs.freezer.addEventListener(`new-chat-${action}`, () => runNewSession(action));
-  }
-  // Agentic memory makes the memory decision in the background, so "skip
-  // memory" stops being a user choice: reduce the affordance to two outcomes —
-  // new session with a saved transcript (save) or without (erase).
-  if (isFeatureEnabled('agentic-memory')) freezerNew()?.setAttribute('no-skip', '');
-
-  // A follower's freezer new-chat routes through the leader (the follower has
-  // no cone / VFS to freeze). wc-tray's LeaderSyncManager relays the request
-  // as this window event; run the same path a local click would run.
-  window.addEventListener(LEADER_RUN_NEW_SESSION_EVENT, (event) => {
-    const action = (event as CustomEvent<Partial<LeaderRunNewSessionDetail>>).detail?.action;
-    if (action === 'save' || action === 'skip' || action === 'erase') runNewSession(action);
-  });
-
-  // By-slug thaw: re-reads the index when the rail hasn't populated yet (URL
-  // routing at boot lands before the first card refresh resolves).
-  const openFrozen = async (slug: string): Promise<void> => {
-    try {
-      const { reader } = await openVfs();
-      let entry = frozenEntries.find((e) => e.filename === slug);
-      if (!entry) {
-        entry = ((await readFreezerEntries(reader)) ?? []).find((e) => e.filename === slug);
-      }
-      // The ARCHIVE is the ground truth: at boot a corrupt index (this deep
-      // link races the rail's self-heal rebuild) used to dead-end here with
-      // a blank shell — thaw straight from the named archive instead.
-      const { messages } = await thawFrozenSession(
-        reader,
-        entry ?? { filename: slug, title: slug, frozenAt: '', messageCount: 0 }
-      );
-      // Set the thread context BEFORE loading messages (mirrors selectScoop's
-      // applyThreadContext-before-load order): loadMessages runs the cone-gated
-      // stale-asset replay, which must see `freezer:…` here — not a stale `cone`
-      // from a prior view — so a thaw can never consume/replay a dropped turn.
-      refs.thread.setAttribute('context', `freezer:${entry?.filename ?? slug}`);
-      // Track the frozen selector id: sessionId for modern entries, filename
-      // for legacy entries that predate the sessionId field. Both reach the
-      // export-service frozen path (the service resolves by either key).
-      currentFrozenSessionId = entry?.sessionId ?? entry?.filename ?? null;
-      getController()?.loadMessages(messages);
-      refs.thread.setAttribute('accent', FREEZER_TINT);
-      // Frost mood: crystallizing shader + ice-blue accent across the frame.
-      applyShellContext(refs, { kind: 'freezer' });
-      refs.inputCard.setAttribute('disabled', '');
-      refs.switcher.removeAttribute('active');
-      clearSelection();
-    } catch (err) {
-      log.error('WC thaw failed', err);
-      // A failed boot deep link must still land somewhere usable: with the
-      // pending URL context already consumed, nothing else selects a scoop —
-      // fall back to the cone rather than a dead empty shell.
-      if (!getSelected()) {
-        const cone = client.getScoops().find((s) => s.isCone);
-        if (cone) selectScoop(cone);
-      }
-    }
-  };
-
-  refs.freezer.addEventListener('freezer-card-select', (event) => {
-    const slug = (event as CustomEvent<{ slug?: string }>).detail?.slug;
-    if (slug) void openFrozen(slug);
-  });
-
-  return {
-    refreshFreezer,
-    openFrozen,
-    getViewedFrozenSessionId: () => currentFrozenSessionId,
-  };
-}
+} from './wc-live-thinking-hydration.js';
+export { scoopColor } from './wc-scoop-color.js';
 
 /** Page-side VFS handles routed through the worker's `VfsRpcHost`. */
 export interface WcPageVfs {
@@ -902,11 +208,6 @@ export function prepareWcShell(app: HTMLElement, floatLabel: string): WcShellBoo
   };
 }
 
-/** Mutable slot for the lazily-wired welcome-flow lick interceptor. */
-export interface WelcomeInterceptHolder {
-  intercept: ((event: LickEvent) => boolean) | null;
-}
-
 /**
  * Page-side VFS factory: the worker owns the (OPFS) filesystem — page reads
  * and writes route through its VfsRpcHost. Opening OPFS from the page would
@@ -954,181 +255,6 @@ function wireWcWelcome(
   });
 }
 
-/** Controller + dip lifecycle over the live agent handle. */
-function createWcController(
-  refs: WcShellRefs,
-  client: OffscreenClient,
-  getSelected: () => RegisteredScoop | null,
-  onIdle?: () => void,
-  welcome?: WelcomeInterceptHolder
-): { controller: WcChatController; agentHandle: ReturnType<OffscreenClient['createAgentHandle']> } {
-  // Dip licks dispatch to the cone as the `inline` sprinkle (the legacy
-  // onDipLick local path), AFTER the welcome-flow interceptor gets first
-  // refusal — onboarding licks must never reach the keyless cone.
-  const dipInstances = new Map<string, DipInstance[]>();
-  void import('../legacy-styles.js')
-    .then(({ loadDipStyles }) => loadDipStyles())
-    .catch(() => undefined);
-  // Keep open dips in step with the WC shell's live theme (the legacy
-  // initTheme path that did this is never called in WC mode).
-  void import('../theme.js')
-    .then(({ watchSprinkleThemeBroadcast }) => watchSprinkleThemeBroadcast())
-    .catch(() => undefined);
-  // Apply persisted custom theme overrides (initTheme is never called in WC mode).
-  void import('../theme-engine.js')
-    .then(({ applyThemeOverrides }) => applyThemeOverrides())
-    .catch(() => undefined);
-  const agentHandle = client.createAgentHandle();
-  // Soundscape cues for the selected scoop's tool lifecycle: tool_use_start →
-  // 'tool-start', tool_result → 'tool-finish'. The cue helper itself gates on
-  // the persisted enable flag, the voice-mode window (`beginVoiceTurn` from
-  // the composer's dictated-submit handler), and the TTS-active flag — so
-  // typed turns stay silent, the wiring just feeds events through.
-  agentHandle.onEvent((event) => {
-    if (event.type !== 'tool_use_start' && event.type !== 'tool_result') return;
-    // A failed tool call earns a 2.6s glower from the focused avatar. This is
-    // the selected scoop's stream, which is exactly whose face is on screen.
-    if (event.type === 'tool_result' && event.isError) refs.switcher.glower();
-    void import('../../speech/soundscape.js')
-      .then(({ playCue }) =>
-        playCue(event.type === 'tool_use_start' ? 'tool-start' : 'tool-finish')
-      )
-      .catch(() => undefined);
-  });
-  const controller = new WcChatController({
-    thread: refs.thread,
-    agent: agentHandle,
-    // Operational telemetry — emit a `formsubmit` beacon per user-initiated
-    // chat send carrying the active scoop name + resolved model id. Resolves
-    // the same way `applyThreadContext` builds the composer meta pill so the
-    // beacon agrees with what the user sees in the UI. Returns null when no
-    // scoop is selected (boot race) so the beacon is skipped rather than
-    // reporting a meaningless empty source.
-    resolveTelemetryContext: () => {
-      const scoop = getSelected();
-      if (!scoop) return null;
-      const scoopName = scoop.isCone ? 'cone' : scoop.name;
-      try {
-        const modelId = scoop.config?.modelId;
-        const model = modelId ? resolveModelById(modelId) : resolveCurrentModel();
-        return { scoopName, model: model.id };
-      } catch {
-        return { scoopName, model: '' };
-      }
-    },
-    // Spoken-reply loop: a turn that began as push-to-talk dictation (the
-    // submit listener marks it) gets its reply read aloud — kokoro once the
-    // chained model download is warm, Web Speech until then. The one-shot
-    // flag is consumed on EVERY turn completion (even reply-less ones, e.g.
-    // the error path) so it can never linger and voice a later typed turn.
-    // The soundscape's voice-mode window also closes here — TTS is flagged
-    // active for the spoken-reply duration so any tail cues are suppressed,
-    // and `endVoiceTurn` runs in a finally so an error path never pins the
-    // voice gate open for later typed turns.
-    onTurnComplete: (message) => {
-      void import('../../speech/voice-reply.js')
-        .then(async ({ consumeVoiceSubmission, speakReplyMarkdown }) => {
-          if (!consumeVoiceSubmission()) return;
-          const { endVoiceTurn, setTtsActive } = await import('../../speech/soundscape.js');
-          try {
-            if (message?.content) {
-              setTtsActive(true);
-              try {
-                await speakReplyMarkdown(message.content);
-              } finally {
-                setTtsActive(false);
-              }
-            }
-          } finally {
-            endVoiceTurn();
-          }
-        })
-        .catch(() => undefined);
-    },
-    onProcessingChange: (processing) => {
-      refs.frame.toggleAttribute('data-processing', processing);
-      refs.inputCard.querySelector('slicc-send-button')?.toggleAttribute('busy', processing);
-      if (!processing) onIdle?.();
-    },
-    // Drive the send button's busy treatment from the turn's phase: `tool`
-    // (spinning ring) while a tool call runs, `thinking` (LLM-wait fill)
-    // otherwise. Only meaningful while the button is `busy`.
-    onBusyPhaseChange: (phase) => {
-      refs.inputCard.querySelector('slicc-send-button')?.setAttribute('phase', phase);
-    },
-    onMessageDisposed: (messageId) => {
-      const instances = dipInstances.get(messageId);
-      if (instances) {
-        disposeDips(instances);
-        dipInstances.delete(messageId);
-      }
-    },
-    onMessageRendered: (message, els) => {
-      const host = els[0];
-      if (!host) return;
-      // Register unconditionally — img-dip placeholders can be pushed
-      // asynchronously into the array after the call returns.
-      dipInstances.set(
-        message.id,
-        hydrateDips(host, (action, data) => {
-          const event: LickEvent = {
-            type: 'sprinkle',
-            sprinkleName: 'inline',
-            timestamp: new Date().toISOString(),
-            body: { action, data },
-          };
-          if (welcome?.intercept?.(event)) return;
-          client.sendSprinkleLick('inline', { action, data });
-        })
-      );
-    },
-    // Route the controller's queued list into the composer's stack. The
-    // component is purely presentational (setMessages replaces the rendered
-    // list); the controller owns enqueue / dismiss / flush-on-consume.
-    onQueuedChange: (items) => {
-      refs.queuedStack.setMessages(items);
-    },
-    onLickBackpressureChange: (notice) => {
-      refs.lickBackpressureNotice.textContent = notice?.text ?? '';
-      refs.lickBackpressureNotice.toggleAttribute('hidden', notice === null);
-    },
-    // Agent-driven `tool_ui` dips (mount approval, USB/serial/HID
-    // pickers) fire their `data-action` clicks through this hook —
-    // forward straight into the worker realm where the request is
-    // registered, otherwise the page-side `toolUIRegistry` (which is
-    // empty in standalone worker mode) silently drops the action and
-    // the tool hangs out its own timeout.
-    onToolUiAction: (requestId, action, data) => {
-      client.sendToolUiAction(requestId, action, data);
-    },
-    // Scoop switch / session reload drops the live-only stack. Route each
-    // dropped id through the SAME backend cancel RPC the `×` dismiss uses
-    // so the orchestrator never delivers a prompt the user implicitly
-    // dropped by navigating away. Best-effort — a transient RPC failure
-    // must not block the local UI from clearing.
-    onQueuedCancel: (messageId) => {
-      const jid = client.selectedScoopJid;
-      if (!jid) return;
-      void client.deleteQueuedMessage(jid, messageId).catch(() => undefined);
-    },
-  });
-  // Local dismiss path: the `×` button on the stack's front card emits a
-  // composed `slicc-queued-remove`; the controller drops the matching item
-  // and re-fires `onQueuedChange` so the stack re-renders. The same id is
-  // also dropped from the orchestrator's queue so the message never reaches
-  // the agent on the next poll.
-  refs.queuedStack.addEventListener('slicc-queued-remove', (event) => {
-    const id = (event as CustomEvent<{ id?: string }>).detail?.id;
-    if (!id) return;
-    controller.removeQueuedMessage(id);
-    const jid = client.selectedScoopJid;
-    if (jid) {
-      void client.deleteQueuedMessage(jid, id).catch(() => undefined);
-    }
-  });
-  return { controller, agentHandle };
-}
-
 export interface AttachWcClientOptions {
   /** Standalone kernel-worker id; enables the sprinkle ops channel. */
   instanceId?: string;
@@ -1140,171 +266,6 @@ export interface AttachWcClientOptions {
     /** Resolved floatbar base label (`sliccstart · live` / `npx · live`). */
     baseFloatLabel?: string;
   };
-}
-
-/**
- * Phase B: wire a connected client into the prepared shell — controller,
- * composer, switcher, workbench, freezer, sprinkles, and nav.
- */
-/**
- * Composer wiring: suggested placeholder (turn-finished hook), the add-menu's
- * real search + staged attachments, submit/stop, and the thinking pill.
- */
-function wireWcComposer(deps: {
-  boot: WcShellBoot;
-  client: OffscreenClient;
-  agentHandle: ReturnType<OffscreenClient['createAgentHandle']>;
-  setRefreshPlaceholder(fn: () => void): void;
-  triggerPlaceholder(): void;
-  openReader(): Promise<WcPageVfs['reader']>;
-  openWriter(): Promise<WcPageVfs['writer']>;
-  log: BootStageLogger;
-}): { getAttachStage(): import('./wc-attach.js').WcAttachmentStage | null } {
-  const { boot, client, agentHandle, openReader, log } = deps;
-  const { refs } = boot;
-  void import('./wc-placeholder.js').then(({ createPlaceholderRefresher }) => {
-    deps.setRefreshPlaceholder(
-      createPlaceholderRefresher({
-        inputCard: refs.inputCard as HTMLElement & { value?: string },
-        getMessages: () => boot.getController()?.getMessages() ?? [],
-        defaultPlaceholder:
-          refs.inputCard.getAttribute('placeholder') ?? 'Ask sliccy, or describe a change…',
-      })
-    );
-  });
-
-  // Hydrate the persisted cone conversation immediately — the worker's
-  // canonical replay (request-scoop-messages on selection) replaces it.
-  // Skipped when the URL deep-links a non-cone context: flashing the cone
-  // history there would be wrong content.
-  // Also skipped for cloud cone followers and cherry embeds — the leader's
-  // snapshot replaces any local history, so showing stale IndexedDB messages
-  // causes flicker.
-  void (async () => {
-    try {
-      if (shouldSkipSessionHydration(boot.wiring.pendingUrlContext, window)) return;
-      const { SessionStore } = await import('../../scoops/chat-session-store.js');
-      const store = new SessionStore();
-      await store.init();
-      const session = await store.load('session-cone');
-      if (session && session.messages.length > 0 && !boot.getSelected()) {
-        boot.getController()?.loadMessages(session.messages);
-        deps.triggerPlaceholder();
-      }
-    } catch (err) {
-      log.warn('WC session hydration failed', err);
-    }
-  })();
-
-  // Add-menu (+): real Files/Skills/Conversations search + staged attachments
-  // (uploads, VFS picks, camera photos, screen captures).
-  let attachStage: import('./wc-attach.js').WcAttachmentStage | null = null;
-  void import('./wc-attach.js')
-    .then(({ wireWcAttach }) => {
-      attachStage = wireWcAttach({
-        inputCard: refs.inputCard as HTMLElement & { value?: string },
-        freezer: refs.freezer,
-        // Camera capture mounts as a compact drop-target box inside the
-        // composer band — mirrors the `<slicc-add-menu>` `.results` geometry.
-        composer: refs.composer,
-        openReader,
-        openWriter: deps.openWriter,
-        listConversations: async () => {
-          const { readSessionsIndex } = await import('../session-freezer.js');
-          const entries = await readSessionsIndex(await openReader());
-          return entries.map((e) => ({
-            id: e.filename,
-            label: e.title,
-            sub: `${e.messageCount} turns`,
-          }));
-        },
-        log,
-      });
-    })
-    .catch((err) => log.error('WC add-menu wiring failed', err));
-
-  // The avatar's expression channels that key off the composer: it makes eye
-  // contact with the input card while it waits on you, attends to each
-  // keystroke with a raised lower lid, and any keystroke wakes it from a drowse.
-  refs.switcher.setAttribute('gaze-target', 'slicc-input-card');
-  refs.inputCard.addEventListener('input', () => {
-    refs.switcher.scrutinize();
-    refs.switcher.wake();
-  });
-
-  refs.inputCard.addEventListener('submit', (event) => {
-    const text = submittedText(event);
-    if (!text) return;
-    // The ball is back in the agent's court — stop waiting on the user, and
-    // rebuild the row NOW. Waiting for the `processing` broadcast (or worse,
-    // the 15s stats poll) leaves the face making eye contact through the whole
-    // send latency, and leaves it stuck there if the send fails before any
-    // status transition lands.
-    boot.wiring.awaitingInput = null;
-    boot.wiring.refreshScoops?.();
-    boot.wiring.notifyScoopStateChanged?.();
-    // Dictated turns (push-to-talk) get their reply spoken back — mark
-    // BEFORE sending so the turn-complete hook sees the flag. The same
-    // flag drives the dictation markers the controller appends to the
-    // sent text so the model knows the input was dictated.
-    const dictation =
-      (event as Event as CustomEvent<{ source?: string }>).detail?.source === 'dictation';
-    if (dictation) {
-      void import('../../speech/voice-reply.js')
-        .then(({ markVoiceSubmission }) => markVoiceSubmission())
-        .catch(() => undefined);
-      // Soundscape: open the voice-mode window and play the message-sent
-      // cue. `beginVoiceTurn` runs BEFORE `playCue('sent')` so the cue's
-      // own voice-turn gate passes; `endVoiceTurn` closes it in the
-      // turn-complete hook above.
-      void import('../../speech/soundscape.js')
-        .then(({ beginVoiceTurn, playCue }) => {
-          beginVoiceTurn();
-          playCue('sent');
-        })
-        .catch(() => undefined);
-    }
-    boot.getController()?.sendUserMessage(text, attachStage?.take(), {
-      dictation,
-      steer: submittedSteer(event),
-    });
-    (refs.inputCard as HTMLElement & { clear?: () => void }).clear?.();
-    // User input is most-recent activity: the addressed scoop gets the eyes
-    // and the text feeds its hover-tooltip summary.
-    const jid = boot.getSelected()?.jid;
-    if (jid) {
-      refs.switcher.setAttribute('attention', jid);
-      boot.wiring.lastActivity.set(jid, text.slice(0, 600));
-    }
-  });
-
-  // The send button morphs into a stop control while a turn is processing.
-  refs.inputCard.addEventListener('stop', () => {
-    if (boot.getController()?.processing) agentHandle.stop();
-  });
-
-  // ArrowUp/ArrowDown in the composer walk the thread's user messages.
-  void import('./wc-history-nav.js')
-    .then(({ wireWcHistoryNav }) =>
-      wireWcHistoryNav({ thread: refs.thread, inputCard: refs.inputCard })
-    )
-    .catch((err) => log.error('WC history nav wiring failed', err));
-
-  // Brain pill: cycle the scoop's thinking level (persisted by the worker).
-  refs.composerMeta.addEventListener('thinking-change', (event) => {
-    if (localStorage.getItem('slicc_locked_effort_level')) return;
-    const metaLevel = (event as CustomEvent<{ thinking?: string }>).detail?.thinking;
-    const level = thinkingLevelForAgent(metaLevel);
-    const effort = effortOverrideForAgent(metaLevel);
-    const selected = boot.getSelected();
-    if (selected && level) {
-      void applyLeaderLocalThinkingChange(client, selected.jid, level, effort).catch((err) =>
-        log.warn('local thinking update failed', err)
-      );
-    }
-  });
-
-  return { getAttachStage: () => attachStage };
 }
 
 /**
@@ -1735,7 +696,8 @@ export function makeSprinkleAttachImage(
   };
 }
 
-// biome-ignore lint/complexity/noExcessiveLinesPerFunction: boot wiring is sequential; extracting more helpers would obscure the order
+/** Sequence the focused live-shell factories over a connected client. */
+// biome-ignore lint/complexity/noExcessiveLinesPerFunction: boot wiring is sequential; focused concerns live in neighboring factories
 export function attachWcClient(
   boot: WcShellBoot,
   client: OffscreenClient,
@@ -1820,137 +782,7 @@ export function attachWcClient(
     openFs: openReader,
     openWriter: async () => (await openVfs()).writer,
     onKernelReady: (fn) => boot.onClientReady(fn),
-    getMonitorDeps: () => ({
-      getScoops: () => client.getScoops(),
-      isProcessing: (jid: string) => client.isProcessing(jid),
-      getCronTasks: async () => {
-        const { getAllCronTasks } = await import('../../scoops/db.js');
-        return getAllCronTasks();
-      },
-      getWebhooks: async () => {
-        const { getAllWebhooks } = await import('../../scoops/db.js');
-        return getAllWebhooks();
-      },
-      getMounts: async () => {
-        const { getAllMountEntries, loadMountHandle } = await import(
-          '../../fs/mount-table-store.js'
-        );
-        const entries = await getAllMountEntries();
-        // Checked fresh on every call (monitor's 5s auto-refresh + manual
-        // "↻ Refresh") — not polled independently. queryPermission() does
-        // NOT require a user gesture (only requestPermission() does), so
-        // this is safe to call here. Mirrors the same check
-        // mount-recovery.ts makes at session-reload time.
-        //
-        // Each local entry below calls loadMountHandle(), which opens (and
-        // closes) its own IndexedDB connection — fine at the 5s monitor
-        // cadence for a normal mount count, but if mount counts grow this
-        // could be batched through a single shared transaction instead.
-        // Deferred: would also change the per-entry error isolation the
-        // try/catch below gives us (one bad handle currently only zeroes
-        // out that one row).
-        return Promise.all(
-          entries.map(async (entry) => {
-            if (entry.descriptor.kind !== 'local') {
-              return entry; // remote backends: no live permission concept, leave `valid` unset
-            }
-            try {
-              const rawHandle = await loadMountHandle(entry.descriptor.idbHandleKey);
-              // queryPermission() is real on Chromium's FileSystemDirectoryHandle but
-              // missing from TS's DOM lib (non-standard across browsers) — one cast.
-              const handle = rawHandle as {
-                queryPermission?: (desc: { mode: string }) => Promise<string>;
-              } | null;
-              const perm = await handle?.queryPermission?.({ mode: 'readwrite' });
-              return { ...entry, valid: perm === 'granted' };
-            } catch {
-              return { ...entry, valid: false };
-            }
-          })
-        );
-      },
-      getMcpServers: async () => {
-        try {
-          const fs = await openReader();
-          const raw = await fs.readFile('/workspace/.mcp/servers.json', { encoding: 'utf-8' });
-          const parsed = JSON.parse(typeof raw === 'string' ? raw : new TextDecoder().decode(raw));
-          return parsed.servers ?? {};
-        } catch {
-          return {};
-        }
-      },
-      getOAuthProviders: () => {
-        try {
-          // getAccounts() already validates the localStorage shape;
-          // route through it (and the account-store OAuth helpers) rather
-          // than re-parsing `slicc_accounts` here, so this stays in sync
-          // with the one canonical reader/writer of that key.
-          const seen = new Set<string>();
-          const entries: { providerId: string; valid?: boolean }[] = [];
-          for (const account of getAccounts()) {
-            if (seen.has(account.providerId)) continue;
-            seen.add(account.providerId);
-            // Only OAuth-flavored providers carry a meaningful valid/invalid
-            // status here — plain API-key accounts get `valid: undefined`
-            // (default/neutral dot), same treatment remote mount backends
-            // get in the mounts section above.
-            if (!getProviderConfig(account.providerId)?.isOAuth) {
-              entries.push({ providerId: account.providerId });
-              continue;
-            }
-            // getOAuthAccountInfo derives status purely from locally-held
-            // fields already on the Account record (loggedOut,
-            // tokenExpiresAt vs. Date.now()) — no token value is read out,
-            // and no network call is made. info === null means no
-            // accessToken at all — e.g. a requiresBaseUrl provider's
-            // base-URL-only row saved before the OAuth popup ever
-            // completed (provider-settings.ts's runOAuthLogin). That's a
-            // real OAuth-flavored account that just never finished
-            // authenticating, so it gets the same red dot as logged-out/
-            // expired rather than falling into the "not an OAuth
-            // provider" neutral case below.
-            const info = getOAuthAccountInfo(account.providerId);
-            const valid = account.loggedOut ? false : info ? !info.expired : false;
-            entries.push({ providerId: account.providerId, valid });
-          }
-          return entries;
-        } catch {
-          return [];
-        }
-      },
-      getSessionStats: async () => client.getSessionStats?.() ?? null,
-      getProcesses: async () => {
-        try {
-          const fs = await openReader();
-          const entries = await fs.readDir('/proc');
-          const procs = [];
-          for (const entry of entries.filter(
-            (e) => e.type === 'directory' && /^\d+$/.test(e.name)
-          )) {
-            try {
-              const stat = await fs.readFile(`/proc/${entry.name}/stat`, {
-                encoding: 'utf-8',
-              });
-              const cmdline = await fs.readFile(`/proc/${entry.name}/cmdline`, {
-                encoding: 'utf-8',
-              });
-              procs.push({
-                pid: parseInt(entry.name, 10),
-                argv: String(cmdline).trim(),
-                status: parseProcStatLine(String(stat)),
-              });
-            } catch {
-              /* process may have exited */
-            }
-          }
-          return procs;
-        } catch {
-          return [];
-        }
-      },
-      getTrayInfo: () => getTrayMonitorInfo(window.localStorage),
-      getConnectedFollowers,
-    }),
+    getMonitorDeps: () => createWcMonitorDeps({ client, openReader, storage: window.localStorage }),
     mountTerminal: (container) => mountWorkbenchTerminal(boot, client, container),
     insertReference: (path: string) => {
       const card = refs.inputCard as HTMLElement & { value: string; focus(): void };
@@ -2261,48 +1093,7 @@ export async function mountWcUiLive(
   // buffer and the worker posts init synchronously, so a late listener would miss
   // a fast boot-time failure.
   if (instanceId) installWorkerStaleAssetReloadListener(instanceId);
-  // Enable the realm sync-fs SW bridge only when a controlling Service Worker
-  // is present to answer `/__slicc/fs-sync/*` (llm-proxy-sw at scope '/', set
-  // up before the UI mounts). Fail-safe: absent → realms use the bounded
-  // snapshot; and the bridge self-guards via a response marker, so a stale
-  // controller can't feed a realm SPA-fallback bytes.
-  const syncFsBridgeEnabled =
-    typeof navigator !== 'undefined' && !!navigator.serviceWorker?.controller;
-  // Per-session nonce naming the sync-fs SW↔responder channel. Minted here (the
-  // page) so it can be handed to BOTH the kernel (spawn init) and the SW
-  // (controller.postMessage) over PRIVATE paths only — realms never receive it,
-  // so they cannot join the channel to steal a capability token or spoof
-  // responses (see sync-fs-wire.ts). A fixed channel name would be joinable by
-  // any same-origin realm — the sandbox escape this closes.
-  const syncFsChannelNonce = syncFsBridgeEnabled ? (crypto.randomUUID() as SyncFsNonce) : undefined;
-  if (syncFsChannelNonce && typeof navigator !== 'undefined' && navigator.serviceWorker) {
-    const sw = navigator.serviceWorker;
-    const nonceMsg: SyncFsNonceMsg = { type: SYNC_FS_NONCE_MSG, nonce: syncFsChannelNonce };
-    const publishNonce = (): void => sw.controller?.postMessage(nonceMsg);
-    publishNonce(); // hand it to the current controller now
-    // Re-publish on a new SW version (controllerchange) and on demand when the
-    // SW lost it (MV3 eviction+respawn asks via sync-fs-need-nonce).
-    sw.addEventListener('controllerchange', publishNonce);
-    sw.addEventListener('message', (event: MessageEvent) => {
-      if ((event.data as SyncFsNeedNonceMsg | undefined)?.type === SYNC_FS_NEED_NONCE_MSG) {
-        publishNonce();
-      }
-    });
-    // Proactive re-arm (fix K): the SW evicts while the tab is idle and
-    // `controllerchange` does NOT re-fire on a respawn, so without this the
-    // first sync-fs op after the user returns eats a cold-start EIO while it
-    // re-requests the nonce. Re-publishing on focus / tab-visible re-arms the
-    // (possibly respawned) SW BEFORE the cone acts, so that op rarely even has
-    // to wait. `addSyncFsNonce` dedupes, so over-publishing is harmless.
-    if (typeof document !== 'undefined') {
-      document.addEventListener('visibilitychange', () => {
-        if (document.visibilityState === 'visible') publishNonce();
-      });
-    }
-    if (typeof window !== 'undefined') {
-      window.addEventListener('focus', publishNonce);
-    }
-  }
+  const { syncFsBridgeEnabled, syncFsChannelNonce } = setupSyncFsBootNonce();
   const host = spawnKernelWorker({
     realCdpTransport,
     instanceId,
